+++ src/inputs.ts
@@ -4,6 +4,7 @@ import * as core from "@actions/core";
 const INPUT_NAME_CONFIG_PATH = "configuration-path";
 const INPUT_NAME_DRY_RUN = "dry-run";
 const INPUT_NAME_REPO_TOKEN = "repo-token";
+const INPUT_NAME_SVGO_OPTIONS = "svgo-options";
 
 const BOOLEAN = "boolean";
 const FALSE = "false";
@@ -12,6 +13,7 @@ const TRUE = "true";
 
 export type RawActionConfig = {
   readonly "dry-run"?: boolean;
+  readonly "svgo-options"?: string;
 }
 
 
@@ -23,6 +25,10 @@ export class ActionConfig {
     this.config = config || { };
   }
 
+  public getSvgoOptionsPath(): string {
+    return this.config["svgo-options"]
+      || core.getInput(INPUT_NAME_SVGO_OPTIONS, { required: false });
+  }
 
   public isDryRun(): boolean {
     const value = (this.config["dry-run"] !== undefined)