+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
+- Configure the SVGO options file. ([#99])
 - Fix `required` value for "configuration-path" input. ([#100])
 - Fix interpretation of "dry-run" input in config file. ([#103])
 
@@ -54,5 +55,6 @@ Versioning].
 [#91]: https://github.com/ericcornelissen/svgo-action/pull/91
 [#92]: https://github.com/ericcornelissen/svgo-action/pull/92
 [#97]: https://github.com/ericcornelissen/svgo-action/pull/97
+[#99]: https://github.com/ericcornelissen/svgo-action/pull/99
 [#100]: https://github.com/ericcornelissen/svgo-action/pull/100
 [#103]: https://github.com/ericcornelissen/svgo-action/pull/103